Add lookupVote functions and key press functionality to proceed through steppers with TextFields
