+++ frontend/src/components/participateWorkflow/ParticipateVoteStepper.jsx
@@ -1,11 +1,10 @@
 import * as React from "react";
+import { useNavigate } from "react-router-dom";
 import { ParticipateContext } from "context/ParticipateContext";
 import { cancelParticipate } from "utils/CancelVote";
 import Stepper from "components/base/Stepper";
 import isMnemonicValid from "utils/IsMnemonicValid";
-import encodeURIMnemonic from "utils/EncodeMnemonic";
-import axios from "axios";
-import { useNavigate } from "react-router-dom";
+import lookupVote from "utils/participateWorkflow/lookupVote";
 
 export default function VerticalLinearStepper() {
 	const [participateInfo, setParticipateInfo] =
@@ -21,7 +20,7 @@ export default function VerticalLinearStepper() {
 		{
 			label: "Enter vote information",
 			description:
-				"Enter the application id and sk to your voting account",
+				"Enter the application id and secret key to your voting account",
 		},
 	];
 
@@ -58,66 +57,7 @@ export default function VerticalLinearStepper() {
 
 	const handleNext = async () => {
 		if (participateInfo.activeStep === 1) {
-			const state = await axios.get(
-				"/api/smartContract/readVoteSmartContractState",
-				{ params: { appId: participateInfo.appId } }
-			);
-
-			const userAccount = await axios.get(
-				"/api/algoAccount/getPublicKey",
-				{
-					params: {
-						mnemonic: encodeURIMnemonic(participateInfo.sk),
-					},
-				}
-			);
-			const addr = userAccount.data.addr;
-
-			const candidates = [];
-			const assetId = state.data.AssetId;
-			const voteBegin = state.data.VoteBegin;
-			const voteEnd = state.data.VoteEnd;
-
-			for (const key of Object.keys(state.data)) {
-				if (
-					key === "AssetId" &&
-					participateInfo.registerOrVote === "vote"
-				) {
-					const assetBalance = await axios.get(
-						"/api/asa/checkAssetBalance",
-						{
-							params: {
-								addr,
-								assetId,
-							},
-						}
-					);
-					if (!assetBalance.data) {
-						alert("You do not hold the vote token for this vote.");
-						return;
-					} else if (assetBalance.data.amount < 1) {
-						alert(
-							"You cannot participate in this vote with a vote token balance of 0."
-						);
-						return;
-					}
-				} else if (
-					!["Creator", "VoteBegin", "VoteEnd", "NumVoters"].includes(
-						key
-					)
-				) {
-					candidates.push(key);
-				}
-			}
-
-			setParticipateInfo({
-				...participateInfo,
-				voteBegin,
-				voteEnd,
-				assetId,
-				candidates,
-				activeStep: participateInfo.activeStep + 1,
-			});
+			lookupVote(participateInfo, setParticipateInfo);
 		} else if (
 			participateInfo.activeStep === 3 ||
 			(participateInfo.activeStep === 2 &&
